coregix/preprocess/registration.py
"""Image registration utilities."""

import subprocess
import sys
from typing import Any, Optional, Sequence, Union


def _require_itk():
    try:
        import itk
    except ImportError as exc:
        raise RuntimeError(
            "itk-elastix is not installed. Install extras with `pip install -e \".[elastix]\"`."
        ) from exc
    return itk


def estimate_elastix_transform(
    fixed_image_path: str,
    moving_image_path: str,
    *,
    parameter_map: Union[str, Sequence[str]] = "rigid",
    parameter_file_paths: Optional[Sequence[str]] = None,
    force_linear_resample: bool = False,
    force_nearest_resample: bool = False,
    fixed_mask_path: Optional[str] = None,
    moving_mask_path: Optional[str] = None,
    log_to_console: bool = False,
) -> Any:
    """Estimate transform parameters that map a moving image to a fixed image.

    Args:
        fixed_image_path: Path to the fixed/reference image.
        moving_image_path: Path to the moving image to be aligned.
        parameter_map: Default elastix parameter map name(s) if no parameter files
            are provided. Typical values include ``"translation"``, ``"rigid"``,
            ``"affine"``, and ``"bspline"``.
        parameter_file_paths: Optional elastix parameter file path(s). If provided,
            these are used instead of ``parameter_map``.
        force_linear_resample: If ``True``, enforce linear final resampling in all
            loaded parameter maps (mimics explicit wrapper settings).
        force_nearest_resample: If ``True``, enforce nearest-neighbor final
            resampling in all loaded parameter maps.
        fixed_mask_path: Optional fixed-image mask path.
        moving_mask_path: Optional moving-image mask path.
        log_to_console: If ``True``, emit elastix logs to stdout.

    Returns:
        ITK transform parameter object produced by elastix.

    Raises:
        RuntimeError: If ``itk-elastix`` is not installed.
    """
    itk = _require_itk()

    fixed = itk.imread(fixed_image_path, itk.F)
    moving = itk.imread(moving_image_path, itk.F)

    parameter_object = itk.ParameterObject.New()
    if parameter_file_paths:
        for path in parameter_file_paths:
            parameter_object.AddParameterMap(parameter_object.ReadParameterFile(path))
    else:
        map_names = [parameter_map] if isinstance(parameter_map, str) else list(parameter_map)
        for map_name in map_names:
            parameter_object.AddParameterMap(parameter_object.GetDefaultParameterMap(map_name))
    if force_linear_resample:
        for idx in range(int(parameter_object.GetNumberOfParameterMaps())):
            parameter_object.SetParameter(idx, "ResampleInterpolator", "FinalLinearInterpolator")
            parameter_object.SetParameter(idx, "DefaultPixelValue", "0")
            parameter_object.SetParameter(idx, "FinalBSplineInterpolationOrder", "1")
    if force_nearest_resample:
        for idx in range(int(parameter_object.GetNumberOfParameterMaps())):
            parameter_object.SetParameter(idx, "ResampleInterpolator", "FinalNearestNeighborInterpolator")
            parameter_object.SetParameter(idx, "DefaultPixelValue", "0")
            parameter_object.SetParameter(idx, "FinalBSplineInterpolationOrder", "0")

    kwargs = {
        "fixed_image": fixed,
        "moving_image": moving,
        "parameter_object": parameter_object,
        "log_to_console": log_to_console,
    }

    if fixed_mask_path:
        kwargs["fixed_mask"] = itk.imread(fixed_mask_path, itk.UC)
    if moving_mask_path:
        kwargs["moving_mask"] = itk.imread(moving_mask_path, itk.UC)

    _, transform_parameter_object = itk.elastix_registration_method(**kwargs)
    return transform_parameter_object


def apply_elastix_transform(
    moving_image_path: str,
    output_image_path: str,
    transform_parameter_object: Any,
    *,
    reference_image_path: Optional[str] = None,
    log_to_console: bool = False,
) -> str:
    """Apply a precomputed elastix transform to an image and write the result.

    Args:
        moving_image_path: Path to the moving image to warp.
        output_image_path: Output path for the transformed image.
        transform_parameter_object: Transform parameter object from
            :func:`estimate_elastix_transform`.
        reference_image_path: Optional raster whose CRS/transform are copied to
            ``output_image_path`` after transformix writes the data.
        log_to_console: If ``True``, emit transformix logs to stdout.

    Returns:
        The written ``output_image_path``.

    Raises:
        RuntimeError: If ``itk-elastix`` is not installed.
    """
    itk = _require_itk()
    moving = itk.imread(moving_image_path, itk.F)
    transformed = itk.transformix_filter(
        moving,
        transform_parameter_object=transform_parameter_object,
        log_to_console=log_to_console,
    )
    itk.imwrite(transformed, output_image_path)
    if reference_image_path:
        import rasterio

        with rasterio.open(reference_image_path) as ref, rasterio.open(output_image_path, "r+") as out:
            out.crs = ref.crs
            out.transform = ref.transform
            if ref.nodata is not None:
                out.nodata = ref.nodata
    return output_image_path


def write_transform_parameter_files(
    transform_parameter_object: Any,
    output_prefix: str,
) -> list[str]:
    """Write an elastix transform parameter object to parameter files."""
    count = int(transform_parameter_object.GetNumberOfParameterMaps())
    parameter_files = [f"{output_prefix}_{idx:03d}.txt" for idx in range(count)]
    itk = _require_itk()
    itk.ParameterObject.WriteParameterFiles(transform_parameter_object, parameter_files)
    return parameter_files


def _apply_elastix_transform_from_parameter_files(
    moving_image_path: str,
    output_image_path: str,
    parameter_files: Sequence[str],
    *,
    reference_image_path: Optional[str] = None,
    log_to_console: bool = False,
) -> str:
    """Apply transformix using serialized parameter files."""
    itk = _require_itk()
    parameter_object = itk.ParameterObject.New()
    itk.ParameterObject.ReadParameterFiles(parameter_object, list(parameter_files))
    return apply_elastix_transform(
        moving_image_path=moving_image_path,
        output_image_path=output_image_path,
        transform_parameter_object=parameter_object,
        reference_image_path=reference_image_path,
        log_to_console=log_to_console,
    )


def apply_elastix_transform_subprocess(
    moving_image_path: str,
    output_image_path: str,
    parameter_files: Sequence[str],
    *,
    reference_image_path: Optional[str] = None,
    log_to_console: bool = False,
) -> str:
    """Apply transformix in a fresh Python subprocess for process isolation."""
    command = [
        sys.executable,
        "-c",
        (
            "from coregix.preprocess.registration import "
            "_apply_elastix_transform_from_parameter_files; "
            "import sys; "
            "_apply_elastix_transform_from_parameter_files("
            "moving_image_path=sys.argv[1], "
            "output_image_path=sys.argv[2], "
            "parameter_files=sys.argv[3:-1], "
            "reference_image_path=None if sys.argv[-1] == '__NONE__' else sys.argv[-1], "
            f"log_to_console={bool(log_to_console)!r})"
        ),
        moving_image_path,
        output_image_path,
        *parameter_files,
        reference_image_path if reference_image_path is not None else "__NONE__",
    ]
    subprocess.run(command, check=True)
    return output_image_path


def run_elastix_registration(
    fixed_image_path: str,
    moving_image_path: str,
    output_image_path: str,
    *,
    parameter_map: Union[str, Sequence[str]] = "rigid",
    parameter_file_paths: Optional[Sequence[str]] = None,
    fixed_mask_path: Optional[str] = None,
    moving_mask_path: Optional[str] = None,
    log_to_console: bool = False,
) -> str:
    """Register a moving image to a fixed image using itk-elastix.

    Args:
        fixed_image_path: Path to the fixed/reference image.
        moving_image_path: Path to the moving image that will be warped.
        output_image_path: Path where the registered image will be written.
        parameter_map: Default elastix parameter map name when no parameter files are provided.
            Common values: ``"rigid"``, ``"affine"``, ``"bspline"``.
        parameter_file_paths: Optional parameter file path(s). When provided, these are used
            instead of ``parameter_map``.
        fixed_mask_path: Optional fixed-image mask path.
        moving_mask_path: Optional moving-image mask path.
        log_to_console: Whether elastix should print logs to stdout.

    Returns:
        The written ``output_image_path``.
    """
    transform_parameter_object = estimate_elastix_transform(
        fixed_image_path=fixed_image_path,
        moving_image_path=moving_image_path,
        parameter_map=parameter_map,
        parameter_file_paths=parameter_file_paths,
        fixed_mask_path=fixed_mask_path,
        moving_mask_path=moving_mask_path,
        log_to_console=log_to_console,
    )
    apply_elastix_transform(
        moving_image_path=moving_image_path,
        output_image_path=output_image_path,
        transform_parameter_object=transform_parameter_object,
        log_to_console=log_to_console,
    )
    return output_image_path


__all__ = [
    "estimate_elastix_transform",
    "apply_elastix_transform",
    "apply_elastix_transform_subprocess",
    "run_elastix_registration",
    "write_transform_parameter_files",
]
