pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]

[tool.setuptools.packages.find]
where = ["."]
include = ["coregix*"]

[tool.setuptools.package-data]
coregix = ["py.typed"]

[project]
name = "coregix"
version = "0.1.0"
description = "Elastix-based geospatial pairwise raster alignment"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Kanoa Lindiwe", email = "cankanoa@gmail.com"},
    {name = "Joseph Emile Honour Percival", email = "ipercival@gmail.com"}
]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "numpy>=2.2",
    "rasterio>=1.4.3",
    "itk-elastix>=0.19.2"
]

[project.urls]
"Source" = "https://github.com/cankanoa/coregix"

[project.scripts]
vhr-align-image-pair = "coregix.cli.align_image_pair:main"
